refactor: move to util secrets functions
