+++ renderer/src/features/mcp-servers/hooks/use-run-custom-server.tsx
@@ -21,11 +21,8 @@ import { Link } from '@tanstack/react-router'
 import { restartClientNotification } from '../lib/restart-client-notification'
 import { trackEvent } from '@/common/lib/analytics'
 import type { FormSchemaLocalMcp } from '../lib/form-schema-local-mcp'
-import {
-  groupSecrets,
-  prepareCreateWorkloadData,
-  saveSecrets,
-} from '../lib/orchestrate-run-local-server'
+import { prepareCreateWorkloadData } from '../lib/orchestrate-run-local-server'
+import { groupMCPDefinedSecrets, saveMCPSecrets } from '@/common/lib/utils'
 
 type InstallServerCheck = (
   data: FormSchemaLocalMcp
@@ -116,7 +113,9 @@ export function useRunCustomServer({
       // We need to know which secrets are new (not from the registry) and which are
       // existing (already stored). This helps us handle the encryption and storage
       // of secrets correctly.
-      const { existingSecrets, newSecrets } = groupSecrets(data.secrets)
+      const { existingSecrets, newSecrets } = groupMCPDefinedSecrets(
+        data.secrets
+      )
 
       // Step 2: Fetch existing secrets & handle naming collisions
       // We need an up-to-date list of secrets so we can handle any existing keys
@@ -136,7 +135,7 @@ export function useRunCustomServer({
       // If there are secrets with values, create them in the secret store first.
       // We need the data returned by the API to pass along with the "run workload" request.
       if (preparedNewSecrets.length > 0) {
-        newlyCreatedSecrets = await saveSecrets(
+        newlyCreatedSecrets = await saveMCPSecrets(
           preparedNewSecrets,
           saveSecret,
           onSecretSuccess,
