+++ renderer/src/features/mcp-servers/hooks/use-update-remote-server.tsx
@@ -20,12 +20,12 @@ import { Button } from '@/common/components/ui/button'
 import { Link } from '@tanstack/react-router'
 import { restartClientNotification } from '../lib/restart-client-notification'
 import { trackEvent } from '@/common/lib/analytics'
+import { prepareUpdateWorkloadData } from '../lib/orchestrate-run-remote-server'
 import {
-  getDefinedSecrets,
-  groupSecrets,
-  prepareUpdateWorkloadData,
-  saveSecrets,
-} from '../lib/orchestrate-run-remote-server'
+  getMCPDefinedSecrets,
+  groupMCPDefinedSecrets,
+  saveMCPSecrets,
+} from '@/common/lib/utils'
 
 type UpdateServerCheck = () => Promise<unknown> | unknown
 
@@ -105,8 +105,9 @@ export function useUpdateServer(
       let newlyCreatedSecrets: SecretsSecretParameter[] = []
 
       // Step 1: Group secrets into new and existing
-      const definedSecrets = getDefinedSecrets(data.secrets)
-      const { existingSecrets, newSecrets } = groupSecrets(definedSecrets)
+      const definedSecrets = getMCPDefinedSecrets(data.secrets)
+      const { existingSecrets, newSecrets } =
+        groupMCPDefinedSecrets(definedSecrets)
 
       // Step 2: Fetch existing secrets & handle naming collisions
       const { data: fetchedSecrets } = await getApiV1BetaSecretsDefaultKeys({
@@ -119,7 +120,7 @@ export function useUpdateServer(
 
       // Step 3: Encrypt secrets
       if (preparedNewSecrets.length > 0) {
-        newlyCreatedSecrets = await saveSecrets(
+        newlyCreatedSecrets = await saveMCPSecrets(
           preparedNewSecrets,
           saveSecret,
           options?.onSecretSuccess || (() => {}),
