+++ renderer/src/features/mcp-servers/lib/orchestrate-run-local-server.tsx
@@ -1,88 +1,13 @@
-import { type UseMutateAsyncFunction } from '@tanstack/react-query'
 import {
   type CoreWorkload,
   type PermissionsOutboundNetworkPermissions,
-  type PostApiV1BetaSecretsDefaultKeysData,
   type SecretsSecretParameter,
   type V1CreateRequest,
-  type V1CreateSecretResponse,
   type V1UpdateRequest,
   type V1ListSecretsResponse,
 } from '@api/types.gen'
-import type { Options } from '@api/client'
-import type { DefinedSecret, PreparedSecret } from '@/common/types/secrets'
 import { getVolumes, mapEnvVars } from '@/common/lib/utils'
 import type { FormSchemaLocalMcp } from './form-schema-local-mcp'
-
-type SaveSecretFn = UseMutateAsyncFunction<
-  V1CreateSecretResponse,
-  string,
-  Options<PostApiV1BetaSecretsDefaultKeysData>,
-  unknown
->
-
-/**
- * Takes all of the secrets from the form and saves them serially to the
- * secret store. Accepts a `toastId`, which it uses to provide feedback on the
- * progress of the operation.
- * // NOTE: We add a short, arbitrary delay to allow the `toast` message that
- * displays progress to show up-to-date progress.
- */
-export async function saveSecrets(
-  secrets: PreparedSecret[],
-  saveSecret: SaveSecretFn,
-  onSecretSuccess: (completedCount: number, secretsCount: number) => void,
-  onSecretError: (
-    error: string,
-    variables: Options<PostApiV1BetaSecretsDefaultKeysData>
-  ) => void
-): Promise<SecretsSecretParameter[]> {
-  const secretsCount: number = secrets.length
-  let completedCount: number = 0
-  const createdSecrets: SecretsSecretParameter[] = []
-
-  for (const { secretStoreKey, target, value } of secrets) {
-    const { key: createdKey } = await saveSecret(
-      {
-        body: { key: secretStoreKey, value },
-      },
-      {
-        onError: (error, variables) => {
-          onSecretError(error, variables)
-        },
-        onSuccess: () => {
-          completedCount++
-          onSecretSuccess(completedCount, secretsCount)
-        },
-      }
-    )
-
-    if (!createdKey) {
-      throw new Error(`Failed to create secret for key "${secretStoreKey}"`)
-    }
-
-    // The arbitrary delay a UX/UI affordance to allow the user to see the progress
-    // of the operation. This is not strictly necessary, but it helps to avoid
-    // confusion when many secrets are being created in quick succession.
-    // The delay is between 100 and 500ms
-    await new Promise((resolve) =>
-      setTimeout(
-        resolve,
-        process.env.NODE_ENV === 'test'
-          ? 0
-          : Math.floor(Math.random() * 401) + 100
-      )
-    )
-    createdSecrets.push({
-      /** The name of the secret in the secret store */
-      name: createdKey,
-      /** The property in the MCP server's config that we are mapping the secret to */
-      target: target,
-    })
-  }
-
-  return createdSecrets
-}
 
 /**
  * Transforms the type specific (e.g. docker vs package manager) data from the
@@ -207,26 +132,6 @@ export function convertWorkloadToFormData(
       image: image || '',
     }
   }
-}
-
-/**
- * A utility function to filter out secrets that are not defined.
- */
-export function getDefinedSecrets(
-  secrets: FormSchemaLocalMcp['secrets']
-): DefinedSecret[] {
-  return secrets.reduce<DefinedSecret[]>((acc, { name, value }) => {
-    if (name && value.secret) {
-      acc.push({
-        name,
-        value: {
-          secret: value.secret,
-          isFromStore: value.isFromStore ?? false,
-        },
-      })
-    }
-    return acc
-  }, [])
 }
 
 // The type of the GET is V1CreateRequest
@@ -354,29 +259,4 @@ export function prepareUpdateWorkloadData(
       : undefined,
     volumes: getVolumes(data.volumes ?? []),
   }
-}
-
-/**
- * Groups secrets into two categories: new secrets (not from the registry) and
- * existing secrets (from the registry). We need this separation to know which
- * secrets need to be encrypted and stored before creating the server workload.
- */
-export function groupSecrets(secrets: DefinedSecret[]): {
-  newSecrets: DefinedSecret[]
-  existingSecrets: DefinedSecret[]
-} {
-  return secrets.reduce<{
-    newSecrets: DefinedSecret[]
-    existingSecrets: DefinedSecret[]
-  }>(
-    (acc, secret) => {
-      if (secret.value.isFromStore) {
-        acc.existingSecrets.push(secret)
-      } else {
-        acc.newSecrets.push(secret)
-      }
-      return acc
-    },
-    { newSecrets: [], existingSecrets: [] }
-  )
 }
