+++ renderer/src/features/mcp-servers/lib/orchestrate-run-remote-server.tsx
@@ -1,86 +1,11 @@
-import { type UseMutateAsyncFunction } from '@tanstack/react-query'
 import {
-  type PostApiV1BetaSecretsDefaultKeysData,
   type SecretsSecretParameter,
   type V1CreateRequest,
-  type V1CreateSecretResponse,
   type V1ListSecretsResponse,
   type V1UpdateRequest,
 } from '@api/types.gen'
-import type { Options } from '@api/client'
 import type { FormSchemaRemoteMcp } from './form-schema-remote-mcp'
-import type { DefinedSecret, PreparedSecret } from '@/common/types/secrets'
 import { mapEnvVars, omit } from '@/common/lib/utils'
-
-type SaveSecretFn = UseMutateAsyncFunction<
-  V1CreateSecretResponse,
-  string,
-  Options<PostApiV1BetaSecretsDefaultKeysData>,
-  unknown
->
-
-/**
- * Takes all of the secrets from the form and saves them serially to the
- * secret store. Accepts a `toastId`, which it uses to provide feedback on the
- * progress of the operation.
- * // NOTE: We add a short, arbitrary delay to allow the `toast` message that
- * displays progress to show up-to-date progress.
- */
-export async function saveSecrets(
-  secrets: PreparedSecret[],
-  saveSecret: SaveSecretFn,
-  onSecretSuccess: (completedCount: number, secretsCount: number) => void,
-  onSecretError: (
-    error: string,
-    variables: Options<PostApiV1BetaSecretsDefaultKeysData>
-  ) => void
-): Promise<SecretsSecretParameter[]> {
-  const secretsCount: number = secrets.length
-  let completedCount: number = 0
-  const createdSecrets: SecretsSecretParameter[] = []
-
-  for (const { secretStoreKey, target, value } of secrets) {
-    const { key: createdKey } = await saveSecret(
-      {
-        body: { key: secretStoreKey, value },
-      },
-      {
-        onError: (error, variables) => {
-          onSecretError(error, variables)
-        },
-        onSuccess: () => {
-          completedCount++
-          onSecretSuccess(completedCount, secretsCount)
-        },
-      }
-    )
-
-    if (!createdKey) {
-      throw new Error(`Failed to create secret for key "${secretStoreKey}"`)
-    }
-
-    // The arbitrary delay a UX/UI affordance to allow the user to see the progress
-    // of the operation. This is not strictly necessary, but it helps to avoid
-    // confusion when many secrets are being created in quick succession.
-    // The delay is between 100 and 500ms
-    await new Promise((resolve) =>
-      setTimeout(
-        resolve,
-        process.env.NODE_ENV === 'test'
-          ? 0
-          : Math.floor(Math.random() * 401) + 100
-      )
-    )
-    createdSecrets.push({
-      /** The name of the secret in the secret store */
-      name: createdKey,
-      /** The property in the MCP server's config that we are mapping the secret to */
-      target: target,
-    })
-  }
-
-  return createdSecrets
-}
 
 /**
  * Combines the registry server definition, the form fields, and the newly
@@ -115,26 +40,6 @@ export function prepareCreateWorkloadData(
 }
 
 /**
- * A utility function to filter out secrets that are not defined.
- */
-export function getDefinedSecrets(
-  secrets: FormSchemaRemoteMcp['secrets']
-): DefinedSecret[] {
-  return secrets.reduce<DefinedSecret[]>((acc, { name, value }) => {
-    if (name && value.secret) {
-      acc.push({
-        name,
-        value: {
-          secret: value.secret,
-          isFromStore: value.isFromStore ?? false,
-        },
-      })
-    }
-    return acc
-  }, [])
-}
-
-/**
  * Transforms form data into an update request object
  */
 export function prepareUpdateWorkloadData(
@@ -162,31 +67,6 @@ export function prepareUpdateWorkloadData(
     env_vars: mapEnvVars(envVars),
     secrets,
   }
-}
-
-/**
- * Groups secrets into two categories: new secrets (not from the registry) and
- * existing secrets (from the registry). We need this separation to know which
- * secrets need to be encrypted and stored before creating the server workload.
- */
-export function groupSecrets(secrets: DefinedSecret[]): {
-  newSecrets: DefinedSecret[]
-  existingSecrets: DefinedSecret[]
-} {
-  return secrets.reduce<{
-    newSecrets: DefinedSecret[]
-    existingSecrets: DefinedSecret[]
-  }>(
-    (acc, secret) => {
-      if (secret.value.isFromStore) {
-        acc.existingSecrets.push(secret)
-      } else {
-        acc.newSecrets.push(secret)
-      }
-      return acc
-    },
-    { newSecrets: [], existingSecrets: [] }
-  )
 }
 
 const getAuthType = (
